Boxplot of the estimated SNP-level ASE

This function is used to showed the estimated SNP-level ASE, i.e. major allele proportion, for one condition sample across all individuals and SNPs after haplotype phasing. The individuals are aligned based on an increasing trend of their median ASE across all snps for a given gene.

plot_ASE(dat, phased = FALSE)

Arguments

dat:
bulk RNA-seq data of a given gene. Must contain variables:

`gene`: character, gene name;

`id`: character, individual identifier;

`snp`: character, the name/chromosome location of the heterzygous genetic variants;

`ref`: numeric, the snp-level read counts for the reference allele if the haplotype phase of the data is unknown, and the snp-level read counts for allele aligned on the same paternal/maternal haplotype for both conditions if haplotype phase is known;

`total`: numeric, snp-level total read counts for both alleles;

phased:
a logical value indicates whether the haplotype phase of the data is known or not. Default is FALSE

Value

The boxplot (ggplot2::geom_boxplot) of the estimated SNP-level ASE across individuals